Add variable for controlling `service.ENABLE_NOTIFY_MAIL` and wire it to `mailer.ENABLED` by default
Without `service.ENABLE_NOTIFY_MAIL` being enabled, the mailer (if enabled) is
only used for important transactional emails, but not for sending
notifications to people about stuff they're watching.

This is not very useful, so we're changing the defaults. If a mailer is
configured, it shall be used for notification emails by default.